+++ src/main/java/com/rlevi/restaurante_backend/dto/PerfilResponseDTO.java
@@ -1,5 +1,6 @@
 package com.rlevi.restaurante_backend.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import com.rlevi.restaurante_backend.model.StatusPedido;
@@ -8,11 +9,11 @@ import lombok.Builder;
 
 @Builder
 public record PerfilResponseDTO(
-        Long pedidoId,
-        String nomeAlimento,
-        Integer quantidade,
-        Double precoUnitario,
-        Double precoTotal,
-        LocalDateTime dataCriacao,
-        StatusPedido status) {
+                Long pedidoId,
+                String nomeAlimento,
+                Integer quantidade,
+                BigDecimal precoUnitario,
+                BigDecimal precoTotal,
+                LocalDateTime dataCriacao,
+                StatusPedido status) {
 }
